+++ src/Configuration/ConfigurationPrefixes.php
@@ -22,7 +22,9 @@ interface ConfigurationPrefixes
     public const CONFIGURATION_PREFIX_PREPAYMENT             = 'prepayment';
     public const CONFIGURATION_PREFIX_TRUSTLY                = 'trustly';
     public const CONFIGURATION_PREFIX_SECURE_INVOICE         = 'secureInvoice';
+    public const CONFIGURATION_PREFIX_OPEN_INVOICE           = 'openInvoice';
     public const CONFIGURATION_PREFIX_APPLE_PAY              = 'applePay';
+    public const CONFIGURATION_PREFIX_BANCONTACT             = 'bancontact';
 
     public const CONFIGURATION_PREFIXES = [
         Handler\PayoneApplePayPaymentHandler::class              => self::CONFIGURATION_PREFIX_APPLE_PAY,
@@ -40,5 +42,7 @@ interface ConfigurationPrefixes
         Handler\PayonePrepaymentPaymentHandler::class            => self::CONFIGURATION_PREFIX_PREPAYMENT,
         Handler\PayoneTrustlyPaymentHandler::class               => self::CONFIGURATION_PREFIX_TRUSTLY,
         Handler\PayoneSecureInvoicePaymentHandler::class         => self::CONFIGURATION_PREFIX_SECURE_INVOICE,
+        Handler\PayoneOpenInvoicePaymentHandler::class           => self::CONFIGURATION_PREFIX_OPEN_INVOICE,
+        Handler\PayoneBancontactPaymentHandler::class            => self::CONFIGURATION_PREFIX_BANCONTACT,
     ];
 }