+++ src/Payone/RequestParameter/Builder/OrderLinesRequestParameterBuilder.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace PayonePayment\Payone\RequestParameter\Builder;
 
 use PayonePayment\Components\Hydrator\LineItemHydrator\LineItemHydratorInterface;
+use PayonePayment\PaymentHandler\PayoneBancontactPaymentHandler;
+use PayonePayment\PaymentHandler\PayoneOpenInvoicePaymentHandler;
 use PayonePayment\PaymentHandler\PayonePayolutionDebitPaymentHandler;
 use PayonePayment\PaymentHandler\PayonePayolutionInstallmentPaymentHandler;
 use PayonePayment\PaymentHandler\PayonePayolutionInvoicingPaymentHandler;
@@ -57,6 +59,8 @@ class OrderLinesRequestParameterBuilder extends AbstractRequestParameterBuilder
             case PayonePayolutionInstallmentPaymentHandler::class:
             case PayonePayolutionInvoicingPaymentHandler::class:
             case PayoneSecureInvoicePaymentHandler::class:
+            case PayoneOpenInvoicePaymentHandler::class:
+            case PayoneBancontactPaymentHandler::class:
                 return true;
         }