+++ src/Payone/RequestParameter/Builder/Refund/RefundRequestParameterBuilder.php
@@ -6,6 +6,7 @@ namespace PayonePayment\Payone\RequestParameter\Builder\Refund;
 
 use PayonePayment\Components\Currency\CurrencyPrecisionInterface;
 use PayonePayment\Installer\CustomFieldInstaller;
+use PayonePayment\PaymentHandler\PayoneDebitPaymentHandler;
 use PayonePayment\Payone\RequestParameter\Builder\AbstractRequestParameterBuilder;
 use PayonePayment\Payone\RequestParameter\Struct\AbstractRequestParameterStruct;
 use PayonePayment\Payone\RequestParameter\Struct\FinancialTransactionStruct;
@@ -44,13 +45,26 @@ class RefundRequestParameterBuilder extends AbstractRequestParameterBuilder
         /** @var CurrencyEntity $currency */
         $currency = $order->getCurrency();
 
-        return [
+        $parameters = [
             'request'        => self::REQUEST_ACTION_DEBIT,
             'txid'           => $customFields[CustomFieldInstaller::TRANSACTION_ID],
             'sequencenumber' => $customFields[CustomFieldInstaller::SEQUENCE_NUMBER] + 1,
             'amount'         => -1 * $this->currencyPrecision->getRoundedTotalAmount((float) $totalAmount, $currency),
             'currency'       => $currency->getIsoCode(),
         ];
+
+        if ($arguments->getPaymentMethod() === PayoneDebitPaymentHandler::class) {
+            $transactionData  = $customFields[CustomFieldInstaller::TRANSACTION_DATA];
+            $firstTransaction = reset($transactionData);
+
+            if (!array_key_exists('request', $firstTransaction) || !array_key_exists('iban', $firstTransaction['request'])) {
+                return $parameters;
+            }
+
+            $parameters['iban'] = $firstTransaction['request']['iban'];
+        }
+
+        return $parameters;
     }
 
     public function supports(AbstractRequestParameterStruct $arguments): bool